添加邀请记录

+++ src/main/java/com/slabs/exchange/common/config/PlatformCoinProperties.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import scala.Int;
 
+import java.math.BigDecimal;
+
 @Data
 @Component
 public class PlatformCoinProperties {
@@ -14,5 +16,5 @@ public class PlatformCoinProperties {
     private String coinName;
 
     @Value("${platform-coin.award-amount}")
-    private Integer awardAmount;
+    private BigDecimal awardAmount;
 }
